Fixed Saving/Restore & Timestamps
Added crude epoch string timestamps to the string's as a quick and dirty
hack to get them in.
Also added a new command /lwtest ( Adds a timestamp & test string at the
end of the loot table )
Fixed Saving / Restoring loot log's ( persist's through reload ui )
NOTE: You must "Clear Logs" At the beginning of a new Raid currently as
the old data is saved through a session. Once the log's have been
cleared (Or at any time after you "/lootlist" you need to "Update Logs"
to "pull" any new changes to the loot list window, this is a temporary
"hackaround" having it automatically update on every /lootlist / on a
timer ( The less timer's we run, the better )

The init. of the the json output lib is done during the first /lootlist,
need to clean this up and find a proper place to initilize it.